build: Move matrix code to main core module
