+++ scheduler-core/src/main/java/com/team01/scheduler/matrix/algorithm/PartialSchedule.java
